+++ framework/iam/bootstrapper.ts
@@ -11,9 +11,8 @@ import BaseOrganizationService from './services/organization.service';
 import BasePermissionService from './services/permission.service';
 import BaseRoleService from './services/role.service';
 import BaseUserService from './services/user.service';
-import { OrganizationService } from './interfaces/organization.service.interface';
 
-const configValidator = {
+export const configValidator = {
   host: string,
   port: number,
   version: optional(string),