+++ packages/express/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @forklaunch/express
 
+## 0.1.30
+
+### Patch Changes
+
+- Updated dependencies
+  - @forklaunch/common@0.1.14
+  - @forklaunch/core@0.2.35
+  - @forklaunch/validator@0.3.13
+
 ## 0.1.29
 
 ### Patch Changes